Preliminary upgrade to the IRC system
