+++ hanyuu/ircbot/irclib/connection.py
@@ -75,7 +75,9 @@ import codecs
 import Queue
 import sqlite3
 import collections
-from . import dcc, utils
+from hanyuu.ircbot.irclib import utils
+from hanyuu.ircbot.irclib import dcc
+
 
 VERSION = 0, 4, 8
 DEBUG = 0
@@ -99,283 +101,6 @@ class IRCError(Exception):
     """Represents an IRC exception."""
     pass
 
-
-class IRC:
-    """Class that handles one or several IRC server connections.
-
-    When an IRC object has been instantiated, it can be used to create
-    Connection objects that represent the IRC connections.  The
-    responsibility of the IRC object is to provide an event-driven
-    framework for the connections and to keep the connections alive.
-    It runs a select loop to poll each connection's TCP socket and
-    hands over the sockets with incoming data for processing by the
-    corresponding connection.
-
-    The methods of most interest for an IRC client writer are server,
-    add_global_handler, remove_global_handler, execute_at,
-    execute_delayed, process_once and process_forever.
-
-    Here is an example:
-
-        irc = irclib.IRC()
-        server = irc.server()
-        server.connect(\"irc.some.where\", 6667, \"my_nickname\")
-        server.privmsg(\"a_nickname\", \"Hi there!\")
-        irc.process_forever()
-
-    This will connect to the IRC server irc.some.where on port 6667
-    using the nickname my_nickname and send the message \"Hi there!\"
-    to the nickname a_nickname.
-    """
-
-    def __init__(self, fn_to_add_socket=None,
-                 fn_to_remove_socket=None,
-                 fn_to_add_timeout=None,
-                 encoding='utf-8'):
-        """Constructor for IRC objects.
-
-        Optional arguments are fn_to_add_socket, fn_to_remove_socket
-        and fn_to_add_timeout.  The first two specify functions that
-        will be called with a socket object as argument when the IRC
-        object wants to be notified (or stop being notified) of data
-        coming on a new socket.  When new data arrives, the method
-        process_data should be called.  Similarly, fn_to_add_timeout
-        is called with a number of seconds (a floating point number)
-        as first argument when the IRC object wants to receive a
-        notification (by calling the process_timeout method).  So, if
-        e.g. the argument is 42.17, the object wants the
-        process_timeout method to be called after 42 seconds and 170
-        milliseconds.
-
-        The three arguments mainly exist to be able to use an external
-        main loop (for example Tkinter's or PyGTK's main app loop)
-        instead of calling the process_forever method.
-
-        An alternative is to just call ServerConnection.process_once()
-        once in a while.
-        """
-
-        if fn_to_add_socket and fn_to_remove_socket:
-            self.fn_to_add_socket = fn_to_add_socket
-            self.fn_to_remove_socket = fn_to_remove_socket
-        else:
-            self.fn_to_add_socket = None
-            self.fn_to_remove_socket = None
-
-        self.fn_to_add_timeout = fn_to_add_timeout
-        self.connections = []
-        self.handlers = {}
-        self.delayed_commands = [] # list of tuples in the format (time, function, arguments)
-        self.encoding = encoding
-        self.add_global_handler("ping", _ping_ponger, -42)
-        
-    def server(self):
-        """Creates and returns a ServerConnection object."""
-
-        c = ServerConnection(self)
-        self.connections.append(c)
-        return c
-
-    def process_data(self, sockets):
-        """Called when there is more data to read on connection sockets.
-
-        Arguments:
-
-            sockets -- A list of socket objects.
-
-        See documentation for IRC.__init__.
-        """
-        for s in sockets:
-            for c in self.connections:
-                if s == c._get_socket():
-                    c.process_data()
-
-    def process_timeout(self):
-        """Called when a timeout notification is due.
-
-        See documentation for IRC.__init__.
-        """
-        t = time.time()
-        while self.delayed_commands:
-            if t >= self.delayed_commands[0][0]:
-                self.delayed_commands[0][1](*self.delayed_commands[0][2])
-                del self.delayed_commands[0]
-            else:
-                break
-
-    def send_once(self):
-        for c in self.connections:
-            try:
-                delta = time.time() - c.last_time
-            except (AttributeError):
-                continue
-            c.last_time = time.time()
-            c.send_time += delta
-            if c.send_time >= 1.3:
-                c.send_time = 0
-                c.sent_bytes = 0
-            
-            while not c.message_queue.empty():
-                if c.sent_bytes <= 2500:
-                    message = c.message_queue.get()
-                    try:
-                        if c.ssl:
-                            c.send_raw_instant(message)
-                        else:
-                            c.send_raw_instant(message)
-                    except (AttributeError):
-                        c.reconnect()
-                    c.sent_bytes += len(message.encode('utf-8'))
-                    if DEBUG:
-                        print("TO SERVER:" + message)
-                else:
-                    break
-
-    def process_once(self, timeout=0):
-        """Process data from connections once.
-
-        Arguments:
-
-            timeout -- How long the select() call should wait if no
-                       data is available.
-
-        This method should be called periodically to check and process
-        incoming data, if there are any.  If that seems boring, look
-        at the process_forever method.
-        """
-        sockets = map(lambda x: x._get_socket(), self.connections)
-        sockets = filter(lambda x: x != None, sockets)
-        if sockets:
-            (i, o, e) = select.select(sockets, [], [], timeout)
-            self.process_data(i)
-        else:
-            time.sleep(timeout)
-        _current_time = time.time()
-        for connection in self.connections:
-            try:
-                _difference = _current_time - connection._last_ping
-            except (AttributeError):
-                continue
-            if (_difference >= 260.0):
-                print("Good morning, client-side ping is here")
-                connection.reconnect("Ping timeout: 260 seconds")
-        self.send_once()
-        self.process_timeout()
-        
-    def process_forever(self, timeout=0.2):
-        """Run an infinite loop, processing data from connections.
-
-        This method repeatedly calls process_once.
-
-        Arguments:
-
-            timeout -- Parameter to pass to process_once.
-        """
-        while 1:
-            self.process_once(timeout)
-
-    def disconnect_all(self, message=""):
-        """Disconnects all connections."""
-        for c in self.connections:
-            c.disconnect(message)
-
-    def add_global_handler(self, event, handler, priority=0):
-        """Adds a global handler function for a specific event type.
-
-        Arguments:
-
-            event -- Event type (a string).  Check the values of the
-            numeric_events dictionary in irclib.py for possible event
-            types.
-
-            handler -- Callback function.
-
-            priority -- A number (the lower number, the higher priority).
-
-        The handler function is called whenever the specified event is
-        triggered in any of the connections.  See documentation for
-        the Event class.
-
-        The handler functions are called in priority order (lowest
-        number is highest priority).  If a handler function returns
-        \"NO MORE\", no more handlers will be called.
-        """
-        if not event in self.handlers:
-            self.handlers[event] = []
-        bisect.insort(self.handlers[event], ((priority, handler)))
-
-    def remove_global_handler(self, event, handler):
-        """Removes a global handler function.
-
-        Arguments:
-
-            event -- Event type (a string).
-
-            handler -- Callback function.
-
-        Returns 1 on success, otherwise 0.
-        """
-        if not event in self.handlers:
-            return 0
-        for h in self.handlers[event]:
-            if handler == h[1]:
-                self.handlers[event].remove(h)
-        return 1
-
-    def execute_at(self, at, function, arguments=()):
-        """Execute a function at a specified time.
-
-        Arguments:
-
-            at -- Execute at this time (standard \"time_t\" time).
-
-            function -- Function to call.
-
-            arguments -- Arguments to give the function.
-        """
-        self.execute_delayed(at-time.time(), function, arguments)
-
-    def execute_delayed(self, delay, function, arguments=()):
-        """Execute a function after a specified time.
-
-        Arguments:
-
-            delay -- How many seconds to wait.
-
-            function -- Function to call.
-
-            arguments -- Arguments to give the function.
-        """
-        bisect.insort(self.delayed_commands, (delay+time.time(), function, arguments))
-        if self.fn_to_add_timeout:
-            self.fn_to_add_timeout(delay)
-
-    def dcc(self, dcctype="chat", dccinfo=(None, None)):
-        """Creates and returns a DCCConnection object.
-
-        Arguments:
-
-            dcctype -- "chat" for DCC CHAT connections or "raw" for
-                       DCC SEND (or other DCC types). If "chat",
-                       incoming data will be split in newline-separated
-                       chunks. If "raw", incoming data is not touched.
-        """
-        c = dcc.DCCConnection(self, dcctype, dccinfo)
-        self.connections.append(c)
-        return c
-
-    def _handle_event(self, connection, event):
-        """[Internal]"""
-        h = self.handlers
-        for handler in h.get("all_events", []) + h.get(event.eventtype(), []):
-            if handler[1](connection, event) == "NO MORE":
-                return
-
-    def _remove_connection(self, connection):
-        """[Internal]"""
-        self.connections.remove(connection)
-        if self.fn_to_remove_socket:
-            self.fn_to_remove_socket(connection._get_socket())
 
 _rfc_1459_command_regexp = re.compile("^(:(?P<prefix>[^ ]+) +)?(?P<command>[^ ]+)( *(?P<argument> .+))?", re.UNICODE)
 
@@ -1032,11 +757,6 @@ class Event(Event):
         """
         arguments = arguments if arguments else []
         super(Event, self).__init__(eventtype, source, target, arguments)
-
-def _ping_ponger(connection, event):
-    """[Internal]"""
-    connection._last_ping = time.time()
-    connection.pong(event.target())
 
 # Numeric table mostly stolen from the Perl IRC module (Net::IRC).
 numeric_events = {
