+++ hanyuu/ircbot/irclib/dcc.py
@@ -2,7 +2,9 @@ from __future__ import unicode_literals
 from __future__ import print_function
 from __future__ import absolute_import
 import socket
-from . import utils, connection
+import hanyuu.ircbot.irclib.utils as utils
+import hanyuu.ircbot.irclib.connection as connection
+
 
 # TODO set this somewhere else?
 DEBUG = False
