+++ hanyuu/ircbot/irclib/session.py
@@ -1,30 +1,302 @@
 """
 A high level session object to the lower level irclib.connection module.
 """
-from . import utils, connection
+from __future__ import unicode_literals
+from __future__ import print_function
+from __future__ import absolute_import
+import hanyuu.ircbot.irclib.utils
+import hanyuu.ircbot.irclib.connection
+import time
+import select
+import bisect
 
 textual_commands = ["privmsg",
                     "pubmsg",
                     "privnotice",
                     "pubnotice"]
+class Session:
+    """Class that handles one or several IRC server connections.
 
-class Session(object):
+    When a Session object has been instantiated, it can be used to create
+    Connection objects that represent the IRC connections.  The
+    responsibility of the Session object is to provide a high-level
+    event-driven framework for the connections and to keep the connections
+    alive. It runs a select loop to poll each connection's TCP socket and
+    hands over the sockets with incoming data for processing by the
+    corresponding connection. It then encapsulates the low level IRC
+    events generated by the Connection objects into higher level
+    versions.
+
+    The methods of most interest for an IRC client writer are server,
+    add_global_handler, remove_global_handler, execute_at,
+    execute_delayed, process_once and process_forever.
+
+    Here is an example:
+
+        irc = irclib.IRC()
+        server = irc.server()
+        server.connect(\"irc.some.where\", 6667, \"my_nickname\")
+        server.privmsg(\"a_nickname\", \"Hi there!\")
+        irc.process_forever()
+
+    This will connect to the IRC server irc.some.where on port 6667
+    using the nickname my_nickname and send the message \"Hi there!\"
+    to the nickname a_nickname.
     """
-    A session object that can join multiple IRC networks and multiple IRC
-    channels with a high level interface.
-    """
-    def __init__(self):
-        super(Session, self).__init__()
-        self.irclib = connection.IRC()
-        
-    def low_level_handler(self, event):
+
+    def __init__(self, fn_to_add_socket=None,
+                 fn_to_remove_socket=None,
+                 fn_to_add_timeout=None,
+                 encoding='utf-8'):
+        """Constructor for IRC objects.
+
+        Optional arguments are fn_to_add_socket, fn_to_remove_socket
+        and fn_to_add_timeout.  The first two specify functions that
+        will be called with a socket object as argument when the IRC
+        object wants to be notified (or stop being notified) of data
+        coming on a new socket.  When new data arrives, the method
+        process_data should be called.  Similarly, fn_to_add_timeout
+        is called with a number of seconds (a floating point number)
+        as first argument when the IRC object wants to receive a
+        notification (by calling the process_timeout method).  So, if
+        e.g. the argument is 42.17, the object wants the
+        process_timeout method to be called after 42 seconds and 170
+        milliseconds.
+
+        The three arguments mainly exist to be able to use an external
+        main loop (for example Tkinter's or PyGTK's main app loop)
+        instead of calling the process_forever method.
+
+        An alternative is to just call ServerConnection.process_once()
+        once in a while.
         """
-        Registered handler for all lowlevel events.
+
+        if fn_to_add_socket and fn_to_remove_socket:
+            self.fn_to_add_socket = fn_to_add_socket
+            self.fn_to_remove_socket = fn_to_remove_socket
+        else:
+            self.fn_to_add_socket = None
+            self.fn_to_remove_socket = None
+
+        self.fn_to_add_timeout = fn_to_add_timeout
+        self.connections = []
+        self.handlers = {}
+        self.delayed_commands = [] # list of tuples in the format (time, function, arguments)
+        self.encoding = encoding
+        self.add_global_handler("ping", self._ping_ponger, -42)
         
-        This should make higher level abstractions of the Event type passed.
-        
+    def server(self):
+        """Creates and returns a ServerConnection object."""
+
+        c = connection.ServerConnection(self)
+        self.connections.append(c)
+        return c
+
+    def process_data(self, sockets):
+        """Called when there is more data to read on connection sockets.
+
+        Arguments:
+
+            sockets -- A list of socket objects.
+
+        See documentation for IRC.__init__.
         """
-        event = HighEvent.from_low_event(event)
+        for s in sockets:
+            for c in self.connections:
+                if s == c._get_socket():
+                    c.process_data()
+
+    def process_timeout(self):
+        """Called when a timeout notification is due.
+
+        See documentation for IRC.__init__.
+        """
+        t = time.time()
+        while self.delayed_commands:
+            if t >= self.delayed_commands[0][0]:
+                self.delayed_commands[0][1](*self.delayed_commands[0][2])
+                del self.delayed_commands[0]
+            else:
+                break
+
+    def send_once(self):
+        for c in self.connections:
+            try:
+                delta = time.time() - c.last_time
+            except (AttributeError):
+                continue
+            c.last_time = time.time()
+            c.send_time += delta
+            if c.send_time >= 1.3:
+                c.send_time = 0
+                c.sent_bytes = 0
+            
+            while not c.message_queue.empty():
+                if c.sent_bytes <= 2500:
+                    message = c.message_queue.get()
+                    try:
+                        if c.ssl:
+                            c.send_raw_instant(message)
+                        else:
+                            c.send_raw_instant(message)
+                    except (AttributeError):
+                        c.reconnect()
+                    c.sent_bytes += len(message.encode('utf-8'))
+                    #if DEBUG:
+                    #    print("TO SERVER:" + message)
+                else:
+                    break
+
+    def process_once(self, timeout=0):
+        """Process data from connections once.
+
+        Arguments:
+
+            timeout -- How long the select() call should wait if no
+                       data is available.
+
+        This method should be called periodically to check and process
+        incoming data, if there are any.  If that seems boring, look
+        at the process_forever method.
+        """
+        sockets = map(lambda x: x._get_socket(), self.connections)
+        sockets = filter(lambda x: x != None, sockets)
+        if sockets:
+            (i, o, e) = select.select(sockets, [], [], timeout)
+            self.process_data(i)
+        else:
+            time.sleep(timeout)
+        _current_time = time.time()
+        for connection in self.connections:
+            try:
+                _difference = _current_time - connection._last_ping
+            except (AttributeError):
+                continue
+            if (_difference >= 260.0):
+                print("Good morning, client-side ping is here")
+                connection.reconnect("Ping timeout: 260 seconds")
+        self.send_once()
+        self.process_timeout()
+        
+    def process_forever(self, timeout=0.2):
+        """Run an infinite loop, processing data from connections.
+
+        This method repeatedly calls process_once.
+
+        Arguments:
+
+            timeout -- Parameter to pass to process_once.
+        """
+        while 1:
+            self.process_once(timeout)
+
+    def disconnect_all(self, message=""):
+        """Disconnects all connections."""
+        for c in self.connections:
+            c.disconnect(message)
+
+    def add_global_handler(self, event, handler, priority=0):
+        """Adds a global handler function for a specific event type.
+
+        Arguments:
+
+            event -- Event type (a string).  Check the values of the
+            numeric_events dictionary in irclib.py for possible event
+            types.
+
+            handler -- Callback function.
+
+            priority -- A number (the lower number, the higher priority).
+
+        The handler function is called whenever the specified event is
+        triggered in any of the connections.  See documentation for
+        the Event class.
+
+        The handler functions are called in priority order (lowest
+        number is highest priority).  If a handler function returns
+        \"NO MORE\", no more handlers will be called.
+        """
+        if not event in self.handlers:
+            self.handlers[event] = []
+        bisect.insort(self.handlers[event], ((priority, handler)))
+
+    def remove_global_handler(self, event, handler):
+        """Removes a global handler function.
+
+        Arguments:
+
+            event -- Event type (a string).
+
+            handler -- Callback function.
+
+        Returns 1 on success, otherwise 0.
+        """
+        if not event in self.handlers:
+            return 0
+        for h in self.handlers[event]:
+            if handler == h[1]:
+                self.handlers[event].remove(h)
+        return 1
+
+    def execute_at(self, at, function, arguments=()):
+        """Execute a function at a specified time.
+
+        Arguments:
+
+            at -- Execute at this time (standard \"time_t\" time).
+
+            function -- Function to call.
+
+            arguments -- Arguments to give the function.
+        """
+        self.execute_delayed(at-time.time(), function, arguments)
+
+    def execute_delayed(self, delay, function, arguments=()):
+        """Execute a function after a specified time.
+
+        Arguments:
+
+            delay -- How many seconds to wait.
+
+            function -- Function to call.
+
+            arguments -- Arguments to give the function.
+        """
+        bisect.insort(self.delayed_commands, (delay+time.time(), function, arguments))
+        if self.fn_to_add_timeout:
+            self.fn_to_add_timeout(delay)
+
+    def dcc(self, dcctype="chat", dccinfo=(None, None)):
+        """Creates and returns a DCCConnection object.
+
+        Arguments:
+
+            dcctype -- "chat" for DCC CHAT connections or "raw" for
+                       DCC SEND (or other DCC types). If "chat",
+                       incoming data will be split in newline-separated
+                       chunks. If "raw", incoming data is not touched.
+        """
+        c = dcc.DCCConnection(self, dcctype, dccinfo)
+        self.connections.append(c)
+        return c
+
+    def _handle_event(self, connection, event):
+        """[Internal]"""
+        h = self.handlers
+        for handler in h.get("all_events", []) + h.get(event.eventtype(), []):
+            if handler[1](connection, event) == "NO MORE":
+                return
+
+    def _remove_connection(self, connection):
+        """[Internal]"""
+        self.connections.remove(connection)
+        if self.fn_to_remove_socket:
+            self.fn_to_remove_socket(connection._get_socket())
+    
+    def _ping_ponger(self, connection, event):
+        """[Internal]"""
+        connection._last_ping = time.time()
+        connection.pong(event.target())
 
 
 class HighEvent(object):
@@ -51,7 +323,20 @@ class HighEvent(object):
                                               *args,
                                               **kwargs)
         
-        if command == 'nick':
+        if command == 'welcome':
+            # We treat this as a "connected" event
+            # The name of the server we are connected to
+            server_name = low_event.source
+            # Our nickname - this might be different than the one we wanted!
+            nickname = Nickname(low_event.target, nickname_only=True)
+            # The welcome message
+            message = low_event.arguments[0]
+            
+            event = creator(nickname, None, message)
+            event.command = 'connect'
+            event.server_name = server_name
+            return event
+        elif command == 'nick':
             # A nickname change.
             old_nickname = Nickname(low_event.source)
             
@@ -69,13 +354,19 @@ class HighEvent(object):
             nickname = Nickname(low_event.source)
             channel = low_event.target
             message = low_event.arguments[0]
-            return creator(nickname, channel, message)
+            event = creator(nickname, channel, message)
+            event.text_command = command
+            event.command = 'text'
+            return event
         elif command in ["privmsg", "privnotice"]:
             # Private message
             # The target is set to our own nickname in privmsg.
             nickname = Nickname(low_event.source)
             message = low_event.arguments[0]
-            return creator(nickname, None, message)
+            event = creator(nickname, None, message)
+            event.text_command = command
+            event.command = 'text'
+            return event
         elif command == 'ctcp':
             # A CTCP to us.
             # Same as privmsg/notice the target is our own nickname
@@ -102,10 +393,22 @@ class HighEvent(object):
             message = low_event.arguments[0]
             
             event = creator(nickname, channel, message)
+            event.text_command = command
+            event.command = 'text'
             return event
         elif command == 'ctcpreply':
-            # A reply to one of our own CTCPs
-            pass
+            # A CTCP reply.
+            # Same as privmsg/notice the target is our own nickname
+            nickname = Nickname(low_event.source)
+            # The irclib splits off the first space delimited word for us.
+            # This is the CTCP command name
+            ctcp = low_event.arguments[0]
+            # The things behind the command are then indexed behind it.
+            message = low_event.arguments[1]
+            
+            event = creator(nickname, None, message)
+            event.ctcp = ctcp
+            return event
         elif command == 'quit':
             # A quit from an user.
             nickname = Nickname(low_event.source)
@@ -165,14 +468,18 @@ class HighEvent(object):
             topic_setter = None
             if command == 'topic':
                 setter = Nickname(low_event.source)            
-            # The argument contains the 
+            # The argument contains the topic string
             # Treat notopic as empty string
             topic = ''
             if not command == 'notopic':
                 topic = low_event.arguments[0]
             event = creator(topic_setter, channel, topic)
+            event.command = 'topic'
             return event
-
+            
+            
+        
+        # The event was not high level: thus, it's raw
         # TODO: Check for missing commands.
         return
